requirements.txt
# ── API Framework ──────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1

# ── ML / Embeddings ────────────────────────────────────────────────────────────
# PyTorch: install with CUDA 12.4 support (matches CUDA 12.7 driver)
# pip install torch==2.6.0 --index-url https://download.pytorch.org/whl/cu124
torch==2.6.0
sentence-transformers==3.0.0
transformers==4.41.0

# ── Vector Store ───────────────────────────────────────────────────────────────
chromadb==0.5.0

# ── Data Processing ────────────────────────────────────────────────────────────
nltk==3.8.1
tqdm==4.66.4
numpy==1.26.4
datasets==2.19.1

# ── HTTP Client (evaluation script) ───────────────────────────────────────────
httpx==0.27.0

# ── Utilities ──────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
